test: restructure suite into concern-focused files
The existing tests/ layout grew organically; test_us.py wasn't really
US-specific, test_llm_usability.py was a grab bag, test_errors.py mixed
error messages with chain-validation assertions, and test_dependencies.py
held only two tests. Redistribute so each file covers one concern.

- tests/test_resolution.py    chain → row: PEP 562 import, name-based
                              resolution, caches, region-name regression,
                              chain validation, empty/no-region paths
- tests/test_discovery.py     dir(), .search(), subtree-scoped list
                              methods, root-only countries/dependencies/
                              subtypes plus their chaining errors
- tests/test_geometry.py      .wkt/.wkb/.geojson, hexwkb/svg stubs,
                              Arrow interop, geometry on empty resolve
- tests/test_errors.py        "Did you mean?" suggestions and empty-
                              result hint content (now pointing at
                              .search() after the Phase 2 migration)
- tests/test_deprecation.py   bracket access shim warnings
- tests/test_configuration.py gains test_overture_version_blocked_on_chained
                              alongside the existing configure() tests

Tests deleted (contents migrated): test_us.py, test_dependencies.py,
test_llm_usability.py, test_geometry_access.py.

Suite count: 91 tests total (89 pass + 2 xfails) vs. 87 pre-refactor.
Net +4 come from added canonical sanity checks (countries == 219,
dependencies == 53, fk.cities == 25, IN.regions == 37) that previously
lived inline in test_access.
