+++ tests/test_errors.py
@@ -1,136 +1,60 @@
+"""Tests for error messages, "Did you mean?" suggestions, and empty-result hints.
+
+When a chain doesn't resolve or a typo leads to a near-miss, `wkls`
+produces suggestions (Levenshtein-based for city names, prefix-based
+for ISO codes) and a tip pointing at ``.search()``. This file covers
+that surface.
+"""
+
+from __future__ import annotations
+
 import pytest
 
 import wkls
 
-
-def test_countries_without_region():
-    """Countries with no regions return an empty DataFrame, not an error."""
-    # Regions now scope to a subtree; countries like FK that have no region
-    # rows simply return zero — no longer a special error case.
-    assert wkls.fk.regions().count() == 0
+# ---------- "Did you mean?" suggestions ----------
 
 
-def test_empty_chain_error():
-    """Test that empty chains raise appropriate errors."""
-    # Create a new Wkl instance without any chain
-    from wkls.core import Wkl
-
-    wkl = Wkl()
-
-    with pytest.raises(ValueError) as exc_info:
-        wkl.resolve()
-    assert "No attributes in the chain" in str(exc_info.value)
-    assert "wkls.<country>" in str(exc_info.value)
-
-
-def test_dependencies_chaining_error():
-    """Test that dependencies() cannot be called on chained objects."""
-    with pytest.raises(ValueError) as exc_info:
-        wkls.us.dependencies()
-    assert "dependencies() can only be called on the root object" in str(exc_info.value)
-    assert "wkls.dependencies()" in str(exc_info.value)
-
-
-def test_countries_chaining_error():
-    """Test that countries() cannot be called on chained objects."""
-    with pytest.raises(ValueError) as exc_info:
-        wkls.us.countries()
-    assert "countries() can only be called on the root object" in str(exc_info.value)
-    assert "wkls.countries()" in str(exc_info.value)
-
-
-def test_regions_at_region_level_returns_self():
-    """regions() at region level returns the region itself — the only region row in that scope."""
-    df = wkls.us.ca.regions().to_arrow_table()
-    assert df.num_rows == 1
-    assert df.column("region")[0].as_py() == "US-CA"
-
-
-def test_regions_past_region_level_raises():
-    """regions() past region level raises (depth 3 has no subtree)."""
-    with pytest.raises(ValueError) as exc_info:
-        wkls.us.ca.sanfrancisco.regions()
-    assert "past region level" in str(exc_info.value)
-
-
-def test_counties_past_region_level_raises():
-    """counties() past region level raises (depth 3 has no subtree)."""
-    with pytest.raises(ValueError) as exc_info:
-        wkls.us.ca.sanfrancisco.counties()
-    assert "past region level" in str(exc_info.value)
-
-
-def test_cities_past_region_level_raises():
-    """cities() past region level raises (depth 3 has no subtree)."""
-    with pytest.raises(ValueError) as exc_info:
-        wkls.us.ca.sanfrancisco.cities()
-    assert "past region level" in str(exc_info.value)
-
-
-def test_subtypes_chaining_error():
-    """Test that subtypes() cannot be called on chained objects."""
-    with pytest.raises(ValueError) as exc_info:
-        wkls.us.subtypes()
-    assert "subtypes() can only be called on the root object" in str(exc_info.value)
-    assert "wkls.subtypes()" in str(exc_info.value)
-
-
-def test_too_many_chained_attributes():
-    """Test that too many chained attributes raise an error."""
-    with pytest.raises(ValueError) as exc_info:
-        wkls.us.ca.sanfrancisco.somethingelse  # noqa: B018
-    assert "Too many chained attributes (max = 3)" in str(exc_info.value)
-
-
-def test_nonexistent_location_errors():
-    """Test errors when trying to access nonexistent locations."""
-    # Nonexistent country
-    try:
-        result = (
-            wkls.zz
-        )  # ZZ is not a valid country code, should return empty DataFrame
-        assert len(result) == 0, "Nonexistent country should return empty DataFrame"
-    except Exception:
-        pass  # Could be various exceptions depending on validation
-
-    # Nonexistent region
-    try:
-        result = (
-            wkls.us.zz
-        )  # ZZ is not a valid state code, should return empty DataFrame
-        assert len(result) == 0, "Nonexistent region should return empty DataFrame"
-    except Exception:
-        pass  # Could be various exceptions depending on validation
-
-    # Nonexistent city with search pattern (this should return empty results)
-    result = wkls.us.ca.search("nonexistentcity")
-    assert result.count() == 0, "Nonexistent city search should return empty DataFrame"
-
-
-def test_geometry_methods_on_empty_results():
-    """Test that geometry methods fail gracefully on empty results."""
-    # Create a chain that will return empty results
-    empty_chain = wkls.us.ca.nonexistentcity
-
-    with pytest.raises(ValueError) as exc_info:
-        empty_chain.wkt()
-    assert "No results found for: us.ca.nonexistentcity" in str(exc_info.value)
-
-
-def test_did_you_mean_suggestions():
-    """Test that typos in location names provide helpful suggestions."""
-    # Typo: "sanfran" instead of "sanfrancisco"
+def test_did_you_mean_city_typo():
+    """Typo in a city name surfaces suggestions via Levenshtein."""
     with pytest.raises(ValueError) as exc_info:
         wkls.us.ca.sanfran.wkt()
-    error_msg = str(exc_info.value)
-    assert "No results found for: us.ca.sanfran" in error_msg
-    assert "Did you mean:" in error_msg
-    assert "sanfrancisco" in error_msg
+    msg = str(exc_info.value)
+    assert "No results found for: us.ca.sanfran" in msg
+    assert "Did you mean:" in msg
+    assert "sanfrancisco" in msg
+
+
+def test_did_you_mean_city_partial_match():
+    """Partial name matches suggest the full name."""
+    with pytest.raises(ValueError) as exc_info:
+        wkls.us.ca.losangel.wkt()
+    msg = str(exc_info.value)
+    assert "Did you mean:" in msg
+    assert "losangeles" in msg
+
+
+def test_did_you_mean_new_york_typo():
+    """Missing-letter typo suggests the correct spelling."""
+    with pytest.raises(ValueError) as exc_info:
+        wkls.us.ny.newyok.wkt()
+    msg = str(exc_info.value)
+    assert "No results found for: us.ny.newyok" in msg
+    assert "Did you mean:" in msg
+    assert "newyork" in msg.lower()
+
+
+def test_did_you_mean_country_level():
+    """Partial country prefixes suggest matching ISO codes."""
+    result = wkls.u
+    repr_str = repr(result)
+    assert "No results found for: u" in repr_str
+    assert "Did you mean:" in repr_str
+    assert "us" in repr_str
 
 
 def test_did_you_mean_region_level():
-    """Test suggestions work for region-level partial codes."""
-    # "c" should suggest codes starting with 'c' like 'ca', 'co', 'ct'
+    """Partial region prefixes suggest matching region ISO suffixes."""
     result = wkls.us.c
     repr_str = repr(result)
     assert "No results found for: us.c" in repr_str
@@ -138,21 +62,8 @@ def test_did_you_mean_region_level():
     assert "ca" in repr_str
 
 
-def test_did_you_mean_country_level():
-    """Test suggestions work for country-level partial codes."""
-    # "u" should suggest codes starting with 'u' like 'ua', 'ug', 'us'
-    result = wkls.u
-    repr_str = repr(result)
-    assert "No results found for: u" in repr_str
-    assert "Did you mean:" in repr_str
-    assert "us" in repr_str
-    # Verify search tip has correct syntax for root level
-    assert "wkls.search('u')" in repr_str
-
-
-def test_suggestions_for_extended_codes():
-    """Test that extended codes suggest the matching prefix."""
-    # "cali" starts with "ca", so suggest "ca"
+def test_did_you_mean_extended_code_suggests_prefix():
+    """Extending a valid prefix (e.g., 'cali') suggests the prefix itself."""
     result = wkls.us.cali
     repr_str = repr(result)
     assert "No results found for: us.cali" in repr_str
@@ -160,76 +71,69 @@ def test_suggestions_for_extended_codes():
     assert "ca" in repr_str
 
 
-def test_no_suggestions_for_unrelated_codes():
-    """Test that completely unrelated codes don't show suggestions."""
-    # "xyz" doesn't match any region code prefix
+def test_did_you_mean_no_region_country():
+    """Suggestions work for 2-level chains on no-region countries like FK."""
+    with pytest.raises(ValueError) as exc_info:
+        wkls.fk.stoney.wkt()
+    msg = str(exc_info.value)
+    assert "No results found for: fk.stoney" in msg
+    assert "Did you mean:" in msg
+    assert "stoneyridge" in msg
+
+
+# ---------- No-suggestions paths ----------
+
+
+def test_no_suggestions_for_unrelated_region_code():
+    """xyz doesn't prefix-match any US region, so no 'Did you mean?'."""
     result = wkls.us.xyz
     repr_str = repr(result)
     assert "No results found for: us.xyz" in repr_str
-    # No suggestions since no code starts with "xyz" and "xyz" doesn't start with any code
     assert "Did you mean:" not in repr_str
-    # But wildcard tip is still shown
+    # But the .search() tip is still shown
     assert "wkls.us.search('xyz')" in repr_str
 
 
-def test_did_you_mean_partial_match():
-    """Test suggestions work for partial name matches."""
-    # "losangel" should suggest "losangeles"
-    with pytest.raises(ValueError) as exc_info:
-        wkls.us.ca.losangel.wkt()
-    error_msg = str(exc_info.value)
-    assert "Did you mean:" in error_msg
-    assert "losangeles" in error_msg
-
-
-def test_no_suggestions_for_completely_wrong_name():
-    """Test that completely wrong names don't show irrelevant suggestions."""
+def test_no_suggestions_for_unrelated_city():
+    """Completely-wrong city names don't surface irrelevant suggestions."""
     with pytest.raises(ValueError) as exc_info:
         wkls.us.ca.xyzabc123.wkt()
-    error_msg = str(exc_info.value)
-    assert "No results found for: us.ca.xyzabc123" in error_msg
-    # Should either have no suggestions or the message shouldn't include irrelevant ones
-    # (cutoff of 0.5 should filter out completely unrelated names)
+    msg = str(exc_info.value)
+    assert "No results found for: us.ca.xyzabc123" in msg
 
 
-def test_did_you_mean_country_without_region():
-    """Test suggestions for countries that don't have regions (2-level chain)."""
-    # Falkland Islands (FK) has no regions, so second level is city directly
-    # Typo: "stoney" instead of "stoneyridge"
-    with pytest.raises(ValueError) as exc_info:
-        wkls.fk.stoney.wkt()
-    error_msg = str(exc_info.value)
-    assert "No results found for: fk.stoney" in error_msg
-    assert "Did you mean:" in error_msg
-    assert "stoneyridge" in error_msg
-
-
-def test_did_you_mean_new_york():
-    """Test suggestions for New York typo at city level."""
-    # Typo: "newyok" (missing 'r') instead of "newyork"
-    with pytest.raises(ValueError) as exc_info:
-        wkls.us.ny.newyok.wkt()
-    error_msg = str(exc_info.value)
-    assert "No results found for: us.ny.newyok" in error_msg
-    assert "Did you mean:" in error_msg
-    # Should suggest newyork or newyorkcity
-    assert "newyork" in error_msg.lower()
-
-
-def test_did_you_mean_no_suggestions_for_region_codes():
-    """Test that region code lookups don't provide fuzzy suggestions."""
-    # Region codes (US-CA, US-NY) are exact matches, no fuzzy matching
-    # When user types wkls.us.xyz, it's treated as region code, not city
+def test_region_codes_do_not_get_fuzzy_suggestions():
+    """Region codes (wkls.us.xyz) use prefix match only, not fuzzy."""
     with pytest.raises(ValueError) as exc_info:
         wkls.us.xyz.wkt()
-    error_msg = str(exc_info.value)
-    assert "No results found for: us.xyz" in error_msg
-    # Region codes are exact - no "Did you mean" for these
+    msg = str(exc_info.value)
+    assert "No results found for: us.xyz" in msg
 
 
-def test_suggestions_in_repr():
-    """Test that suggestions appear in DataFrame repr for empty results."""
-    # Access a non-existent city - should return empty DataFrame with suggestions in repr
+# ---------- Empty-result hint shows .search() ----------
+
+
+def test_error_hint_uses_search_not_brackets():
+    """Empty-result hint points at .search(), not bracket wildcards."""
+    repr_str = repr(wkls.us.ca.nonexistentcity)
+    assert "wkls.us.ca.search('nonexistentcity')" in repr_str
+    assert "['%" not in repr_str
+
+
+def test_error_hint_at_root_uses_search():
+    """Root-level empty result hints at wkls.search()."""
+    repr_str = repr(wkls.zz)
+    assert "wkls.search('zz')" in repr_str
+
+
+def test_error_hint_at_country_level_uses_search():
+    """Country-level empty result hints at wkls.<country>.search()."""
+    repr_str = repr(wkls.us.xyz)
+    assert "wkls.us.search('xyz')" in repr_str
+
+
+def test_suggestions_show_in_empty_repr():
+    """Non-raising empty chains include suggestions + search tip in repr."""
     result = wkls.us.ca.sanfran
     repr_str = repr(result)
     assert "No results found for: us.ca.sanfran" in repr_str
@@ -238,11 +142,8 @@ def test_suggestions_in_repr():
     assert "wkls.us.ca.search('sanfran')" in repr_str
 
 
-def test_suggestions_in_repr_bracket_access(ignore_bracket_deprecation):
-    """Test that suggestions appear when using bracket access."""
-    # Access a non-existent country using bracket syntax on a Wkl instance.
-    # Bracket access is deprecated but must still work as a shim — that's
-    # exactly what this test guards against regressing.
+def test_suggestions_via_Wkl_bracket_access(ignore_bracket_deprecation):
+    """Bracket access on a Wkl instance surfaces the same hint."""
     from wkls import Wkl
 
     result = Wkl()["uss"]
@@ -253,53 +154,11 @@ def test_suggestions_in_repr_bracket_access(ignore_bracket_deprecation):
     assert "wkls.search('uss')" in repr_str
 
 
-def test_chainable_dataframe_error_propagation():
-    """Test that ChainableDataFrame properly propagates errors."""
-    # Get a valid DataFrame first
-    us_data = wkls.us
-
-    # countries() should fail on chained data
-    with pytest.raises(ValueError) as exc_info:
-        us_data.countries()
-    assert "countries() can only be called on the root object" in str(exc_info.value)
-
-    # regions() on a region-level chain returns the region itself.
-    ca_data = wkls.us.ca
-    assert ca_data.regions().count() == 1
+# ---------- Error propagation through ChainableDataFrame ----------
 
 
-def test_version_attribute():
-    """Test that __version__ is accessible and returns a string."""
-    assert hasattr(wkls, "__version__")
-    assert isinstance(wkls.__version__, str)
-    assert len(wkls.__version__) > 0
-
-
-def test_dunder_attributes_raise_attribute_error():
-    """Test that unset dunder attributes raise AttributeError, not empty DataFrames."""
-    with pytest.raises(AttributeError):
-        wkls.__nonexistent__  # noqa: B018
-
-    with pytest.raises(AttributeError):
-        wkls._private_attr  # noqa: B018
-
-
-if __name__ == "__main__":
-    test_empty_chain_error()
-    test_countries_chaining_error()
-    test_regions_at_region_level_returns_self()
-    test_regions_past_region_level_raises()
-    test_counties_past_region_level_raises()
-    test_cities_past_region_level_raises()
-    test_subtypes_chaining_error()
-    test_too_many_chained_attributes()
-    test_nonexistent_location_errors()
-    test_geometry_methods_on_empty_results()
-    test_did_you_mean_suggestions()
-    test_did_you_mean_partial_match()
-    test_no_suggestions_for_completely_wrong_name()
-    test_did_you_mean_country_without_region()
-    test_did_you_mean_new_york()
-    test_did_you_mean_no_suggestions_for_region_codes()
-    test_chainable_dataframe_error_propagation()
-    print("All error handling tests passed!")
+def test_chainable_df_propagates_root_only_errors():
+    """Root-only methods raise when called through ChainableDataFrame."""
+    us_df = wkls.us
+    with pytest.raises(ValueError, match="root object"):
+        us_df.countries()
